Fix small typo, add library imports in examples (#137)
* Fix small typo, add library imports in examples

Thank you for this amazing project 👍 

I just spotted a small typo while working through the Getting Started materials and thought the most direct way to communicate this is probably via a PR. I also added imports for pandas, streamlit & numpy the first time they are referenced in the examples. I believe this adds clarity for newcomers.

* Remove reference to non-existent prior snippet

In the Getting Started materials when introducing the "Magic" approach a prior code snippet is mentioned which, in the current state of the docs, doesn't exit.