Add a tip mentioning Python module invocation of Streamlit
Along with the upcoming knowledge base articles dedicated to describing
how to configure Streamlit with popular IDEs/editors, it may be useful
to also add a small snippet briefly mentioning Python module invocation
of Streamlit. This PR does this in the main-concepts doc.

It feels a bit weird to me that we have two tips back to back, so
suggestions on a better place to put this are appreciated :)


Former-commit-id: dfcf781afc8e440bc1699343959f9156b6eecfa5