Added documentation for in-app options menu (use and configuration) (#724)
* Update Configuration with latest config options

Config options updated with latest versions and shortened to 79 character per line to prevent side-scrolling in markdown code blocks.

* Corrected showErrorDetails description

Corrected the description of `showErrorDetails` which does not currently hide the traceback if set to false.

* New page to document in-app options menu

Created new page to document the options menu available in the upper-right corner of a running app.

* Moved configuration documentation to api section

Completed draft of app menu documentation and moved related configuration documentation within "Configuration and management" with the API reference directory. URLs throughout the documentation were updated and redirects were added.

* Revert "Moved configuration documentation to api section"

This reverts commit 3870d77f26479e9d8a2c69a078ae2cc0b3221a74 [formerly 79086b3f2fc6baa2510472a8c727fc40055cba76].

* Additional content for app menu documentation

Draft documentation of app menu options.

* Move App menu documentation to Advanced features

Moved App menu documentation from `/api-reference/configuration-and-management` to `/advanced-features`.

* Removed outdated KB articles

How to record a screencast and how to hide the hamburger menu are not part of the Streamlit library documentation. KB articles for these topics are removed with this commit.

* Improve alignment of images

* Swap clear and deploy source images

* Add entire config file with Collapse component

* Fix broken link to config page in Customize the menu

* Link to session state explanatory docs instead of API

* Typo: Be sure -> Ensure

* Reverted clear and deploy images

Developer options menu is grey with a white highlight; reverted images to show white on hover.

* Update config option definitions

Updated config option definitions to match streamlit/streamlit#6917

* Image position: mobile compatibility

Updated margin on images so mobile display is also pretty. Using auto margin on small images not expected to fill the screen and left+right equal margin for images expected to be full-width.

* Update print option description

Remove reference to limited/experimental and specify that it just uses the browser's print=to-pdf function.

---------

Co-authored-by: snehankekre <snehan.minerva@gmail.com>
Former-commit-id: 43f01d75254fcd923f2f80e23307aa4426f4240c