Update images to remove hamburger menu (#730)
* Updated images to remove hamburger menu

In-app menu was change from a hamburger to a kebab. This commit updated the images on the App Menu page of the docs.

* Remove references to hamburger menu

This commit replaces any reference to "hamburger" or "☰" in the docs. Older references to a hamburger menu in the Community Cloud admin panel were also removed as that icon was previously changed to a vertical ellipse also.

* Change "menu" to "icon" for click instructions

Update wording to indicate "click an icon" to "open a menu" (as opposed to "click a menu").